Refactor dos arquivos modularizando em classes e refactor da estrutura de pastas
