+++ app/static/css/style.css
@@ -1,0 +1,68 @@
+body {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 80vh;
+    margin: 0;
+    font-family: 'Poppins', sans-serif;
+}
+
+h1 {
+    text-align: center;
+}
+
+img {
+    border-radius: 50%;
+    width: 240px;
+    height: 240px;
+    object-fit: cover;
+}
+
+.input-container {
+    text-align: center;
+    margin-top: 20px;
+    width: 100%;
+    max-width: 600px;
+}
+
+input[type="text"] {
+    padding: 10px;
+    font-family: inherit;
+    width: 100%;
+    border: 0;
+    border-bottom: 2px solid #9b9b9b;
+    outline: 0;
+    font-size: 1.3rem;
+    color: #525252;
+    padding: 7px 0;
+    background: transparent;
+    transition: border-color 0.2s;
+    text-align: center;
+}
+
+input[type="text"]::placeholder {
+    color: #525252;
+}
+
+input[type="text"]:placeholder-shown {
+    font-size: 1.3rem;
+    cursor: text;
+    top: 20px;
+}
+
+input[type="text"]:focus {
+    display: block;
+    transition: 0.2s;
+    color: #3f526d;
+}
+
+.output {
+    margin-top: 20px;
+    font-style: italic;
+}
+
+.tips {
+    margin-top: 40px;
+
+}
