+++ app/static/js/scripts.js
@@ -1,0 +1,4 @@
+// Scripts
+document.addEventListener("DOMContentLoaded", function () {
+    console.log("Script carregado e funcionando!");
+});
