--- templates/index.html
@@ -1,166 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Lisa</title>
-    <style>
-        body {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            margin: 0;
-            margin-top: 40px;
-            margin-bottom: 80px;
-            font-family: 'Poppins', sans-serif;
-        }
-
-        h1 {
-            text-align: center;
-            margin-bottom: 0;
-        }
-
-        .intro {
-            margin-bottom: 20px;
-        }
-
-        img {
-            border-radius: 50%;
-            width: 240px;
-            height: 240px;
-            object-fit: cover;
-        }
-
-        .input-container {
-            text-align: center;
-            margin-top: 20px;
-            width: 100%;
-            max-width: 600px;
-        }
-
-        input[type="text"] {
-            padding: 10px;
-            font-family: inherit;
-            width: 100%;
-            border: 0;
-            border-bottom: 2px solid #9b9b9b;
-            outline: 0;
-            font-size: 1.3rem;
-            color: #525252;
-            padding: 7px 0;
-            background: transparent;
-            transition: border-color 0.2s;
-            text-align: center;
-        }
-
-        input[type="text"]::placeholder {
-            color: #525252;
-        }
-
-        input[type="text"]:placeholder-shown {
-            font-size: 1.3rem;
-            cursor: text;
-            top: 20px;
-        }
-
-        input[type="text"]:focus {
-            display: block;
-            transition: 0.2s;
-            color: #3f526d;
-        }
-
-        .output {
-            margin-top: 20px;
-            font-style: italic;
-        }
-
-        .tips {
-            margin-top: 40px;
-
-        }
-
-        .roteiro-gerado {
-            margin-top: 20px;
-        }
-
-        .roteiro-gerado h2 {
-            color: #3f526d;
-            padding-bottom: 20px;
-        }
-
-        .secao .detalhes {
-            margin-top: 0;
-            line-height: 21px;
-        }
-
-        .secao h3 {
-            margin-bottom: 5px;
-        }
-
-        .secao strong {
-            color: #5c5c5c;
-        }
-
-        footer {
-            margin-top: 30px;
-            text-align: center;
-        }
-    </style>
-</head>
-
-<body>
-    <div class="intro">
-        <h1>Olá, eu sou Lisa</h1>
-        <small>Uma assistente virtual na criação de roteiros para retrospectivas.</small>
-    </div>
-
-    <img src="{{ url_for('static', filename='lisa-avatar-320.jpg') }}" alt="Lisa">
-
-
-    {% if roteiro %}
-    <div class="roteiro-gerado">
-        <h2>Ok, aqui está minha sugestão de roteiro:</h2>
-        {% for secao, (atividade, detalhes) in roteiro.items() %}
-        <div class="secao">
-            <h3>{{ secao }}</h3>
-            <p class="detalhes">
-                <strong>Duração:</strong> {{ detalhes['duracao'] }}<br>
-                <strong>Formação:</strong> {{ detalhes['formacao'] }}<br>
-                <strong>Atividade:</strong> {{ atividade }}<br>
-                <strong>Como funciona?</strong> {{ detalhes['descricao'] }}
-            </p>
-        </div>
-        {% endfor %}
-    </div>
-    {% else %}
-    <div class="input-container">
-        <form method="POST" id="inputForm">
-            <input type="text" name="user_input" placeholder="Me conta, sobre o que vai ser a Retrospectiva hoje?"
-                onkeydown="if(event.key === 'Enter'){this.form.submit(); return false;}">
-        </form>
-    </div>
-    <div class="tips">
-        <h4>Dicas:</h4>
-        <ul>
-            <li>Diga o nome do time onde será feita a Retrospectiva.</li>
-            <li>Conte para Lisa o tipo de problema que o time está enfrentando.</li>
-            <li>Fale sobre a duração esperada de timebox.</li>
-        </ul>
-    </div>
-    {% endif %}
-
-    <footer>
-        <hr>
-        <small>
-            Desenvolvido com ❤️ por <a href="https://www.linkedin.com/in/felipefernand10/" target="_blank">Felipe
-                Fernandes</a> 😊<br>
-            Gostou? Tem alguma idéia e quer contribuir com esse projeto? <a
-                href="https://github.com/felipefernandes/lisa" target="_blank">Github</a>.
-        </small>
-    </footer>
-
-</body>
-
-</html>
